# apps/api/docs/openapi.yaml
openapi: 3.1.0
...
info:
  title: Ownpace — managed API
  version: 0.1.0-rc.1
  summary: Migrate mail, calendar, contacts and files off US cloud, on your own schedule.
  description: |
    The multi-tenant managed edition. Every request is scoped to one tenant by
    the bearer token; the database enforces the same boundary again with
    row-level security, so a token for tenant A cannot read tenant B even if a
    route forgets to filter.

    **Two ideas run through the whole surface and explain most of its shape:**

    - *Nothing consequential happens without a person.* Deleting on the target,
      finishing a migration and moving billing money are gated behind explicit
      evidence, an owner/admin role, or both. Endpoints that could destroy data
      refuse by default and say what is missing.
    - *A failure is reported as itself.* Routes return the reason verbatim
      rather than an empty success. "I found nothing" and "I could not look"
      are different answers and never share a status code.

    Long-running work (sync, cutover, verification, applying a deletion) is
    enqueued, not performed in the request: those endpoints answer `202` with
    an identifier, and a matching read endpoint polls the outcome.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: Ownpace
    url: https://github.com/Robbes/Ownpace
servers:
  - url: https://{host}
...
    variables:
      host:
        default: localhost:3001
...
security:
  - bearerAuth: []
...
paths:
...
  /api/connections:
    get:
      tags: [Connections]
      summary: The tenant's source and target connections
      description: |
        Connection rows have existed since the baseline but were never shown:
        creating a mapping silently inserts two and that was the last anyone
        heard of them. This lists what exists, what each is for, and how many
        mailboxes depend on it. Secrets never come back out — names, kinds and
        states only.
      responses:
        '200':
          description: The connections.
          content:
            application/json:
              schema:
                type: object
                properties:
                  connections:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string, format: uuid }
                        role: { type: string, enum: [source, target] }
                        kind: { type: string }
                        displayName: { type: string }
                        status: { type: string, enum: [connected, error, revoked] }
                        createdAt: { type: string, format: date-time }
                        usedByMailboxes:
                          type: integer
                          description: How many mailboxes depend on this connection.
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Connections]
      summary: Add a connection on its own, without creating a mapping
      description: |
        Which fields a provider needs comes from the shared credential
        descriptor, so a form and this route cannot disagree about them; the
        stored config and encrypted credential record are built by the same
        functions the create-mapping route uses, so what is saved here is what
        a sync pass reads. Probed before it is stored, and stored either way
        with the outcome on `status` — a credential that does not work yet is
        worth keeping (somebody is mid-setup) but must not look healthy.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role, type, displayName, values]
              properties:
                role: { type: string, enum: [source, target] }
                type:
                  type: string
                  description: A wizard source or target type, e.g. box or webdav.
                displayName: { type: string }
                values:
                  type: object
                  additionalProperties: { type: string }
                  description: Field key to what the person typed. Secrets included; never echoed back.
      responses:
        '201':
          description: Created, with what the probe found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string, format: uuid }
                  ok: { type: boolean }
                  detail: { type: string }
                  reason: { type: string }
        '400': { description: Unknown type, a missing required field, or a value the schema refuses. }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/connections/{id}:
    delete:
      tags: [Connections]
      summary: Delete a connection nothing depends on
      description: |
        REFUSES while any mailbox uses it, naming which migrations. That
        refusal is the feature: `mailbox.connection_id` cascades on delete and
        `item` hangs off the mailboxes, so deleting a connection in use would
        not fail loudly — it would take the mailboxes and the whole migration
        ledger with it, and the next pass would re-copy everything as though it
        had never run.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204': { description: Deleted. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { description: No such connection for this tenant. }
        '409':
          description: Still in use; the reason names how many and which migrations.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/connections/{id}/credentials:
    put:
      tags: [Connections]
      summary: Replace a connection's credentials in place
      description: |
        Credentials expire — a rotated Box secret, a revoked Google refresh
        token — and without this the only repair was a new connection and a new
        mapping, abandoning the ledger history that made the old one useful.
        The row keeps its id, so every mapping pointing at it keeps working.
        The new credentials are PROBED BEFORE they replace the old ones, and a
        failure leaves the old ones in place (`rotated: false`): turning a
        working connection into a broken one because somebody pasted the wrong
        value is worse than refusing. The stored config is deliberately not
        touched — rotation replaces a secret, not where a migration is rooted.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [values]
              properties:
                values:
                  type: object
                  additionalProperties: { type: string }
      responses:
        '200':
          description: What the probe found, and whether the credentials were replaced.
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  detail: { type: string }
                  reason: { type: string }
                  rotated:
                    type: boolean
                    description: False when the probe failed and the old credentials were kept.
        '400': { description: A missing required field, or a value the schema refuses. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { description: No such connection for this tenant. }
  /api/connections/{id}/test:
    post:
      tags: [Connections]
      summary: Probe a stored connection now
      description: |
        Runs the same read-only probe the wizard's Test connections button
        uses, through the builders a sync pass would use, against the stored
        credentials — so a credential that has expired says so here instead of
        in a failing pass. The provider's refusal comes back verbatim, and the
        outcome is recorded on the connection's status.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: What the provider answered. A refusal is an answer, not an error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  detail: { type: string }
                  reason: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { description: No such connection for this tenant. }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        A tenant-scoped JWT. Local deployments sign it with `JWT_SECRET`;
        managed SSO validates against a JWKS instead.

        The token's subject must be a MEMBER of the tenant it names — an
        arbitrary subject with a valid signature is refused. Roles (`owner`,
        `admin`, `member`, `viewer`) gate the routes marked owner/admin above,
        for reads as well as writes on billing.
...
  responses:
...
    Unauthorized:
      description: Missing, invalid, or non-member token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Error:
      type: object
      description: Every refusal has this shape. `message` is meant to be shown to a person as-is.
      properties:
        error: { type: string, examples: ['Conflict'] }
        message:
          type: string
          examples: ['Mapping is paused — review the discovery counts and start it first (POST /start).']
      required: [error, message]
...